perf(subscriptions): Add script for load testing snuba subscriptions
This script allows us to bulk create subscriptions in snuba. It creates them with random
permutations of various parameters.  We could potentially add some more items to `queries` to get
some queries with more terms going.

We create all the subscriptions with a fake type. This means that the `QuerySubscriptionConsumer`
will receive the updates, load the `QuerySubscription`, find no handler and just exit. So we're only
testing load on the Snuba/Clickhouse side for now.

Right now this is just on Sentry, but we can adapt it to add subscriptions to other random projects
too. These will have no effect on anything, so it's safe to do so.

When we're finished with the load test, or if we just need to stop it, we can run
`delete_fake_subscriptions` and it'll remove all the subscriptions that were created with this
script.

Probably no need to commit this to master.